refactor: transform movie to module
