+++ src/app/app.routes.ts
@@ -3,7 +3,6 @@ import { Routes } from '@angular/router';
 export const routes: Routes = [
   {
     path: '',
-    loadComponent: () =>
-      import('./features/movie/pages/movies/movies.component'),
+    loadComponent: () => import('./features/movie/movies.module'),
   },
 ];
