--- src/app/features/movie/pages/movies/movies.component.html
@@ -1,25 +1,0 @@
-<div class="w-full flex flex-col items-center justify-center bg-cover bg-center bg-no-repeat h-[28rem]"
-  [ngClass]="{ 'bg-gray-800': !currentMovie()}"
-  [ngStyle]="{ 'background-image': 'url(' + currentMovie()?.backdrop + ')' }">
-  @if (currentMovie()) {
-  <div class="p-4 w-9/10 sm:w-4/5 md:w-3/5 lg:w-2/5 flex flex-col gap-2 lg:gap-4">
-    <h1 class="text-white font-bold text-2xl">{{ currentMovie()?.title }}</h1>
-    <div class="flex flex-col gap-1 lg:gap-2">
-      <p class="text-white mt-4">{{ currentMovie()?.overview }}</p>
-      <p class="text-amber-300 mt-4">Rating: {{ currentMovie()?.score }}/10</p>
-    </div>
-    <movie-search (search)="search($event)"></movie-search>
-  </div>
-  } @else {
-  <div class="w-9/10 sm:w-4/5 md:w-3/5 lg:w-2/5">
-    <movie-search (search)="search($event)"></movie-search>
-  </div>
-  }
-</div>
-@if (movieLists()) {
-<div class="flex items-center justify-center mt-8">
-  <movie-list [list]="movieLists()" (select)="selectMovie($event)"></movie-list>
-</div>
-} @else {
-<p>No movies available</p>
-}
